server/requirements.txt
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
protobuf==3.18.0
six==1.16.0
Werkzeug==2.0.1